# api/swagger.yaml
basePath: /web
definitions:
...
  web.CategoriesResponse:
    properties:
      categories:
        items:
          type: string
        type: array
    type: object
...
  web.CreateTransactionRequest:
    properties:
      amount:
        example: 12.5
        type: number
      category:
        example: Food
        type: string
      date:
        example: "2026-05-21"
        type: string
      description:
        example: lunch
        type: string
      type:
        example: Expense
        type: string
    type: object
  web.DeleteTransactionRequest:
    properties:
      id:
        type: integer
    type: object
...
  web.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  web.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  web.StatsResponse:
    properties:
      balance:
        type: number
      totalExpenses:
        type: number
      totalIncome:
        type: number
      totalTransactions:
        type: integer
    type: object
  web.TransactionDTO:
    properties:
      amount:
        type: number
      category:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: integer
      type:
        type: string
    type: object
  web.TransactionsResponse:
    properties:
      count:
        type: integer
      transactions:
        items:
          $ref: '#/definitions/web.TransactionDTO'
        type: array
    type: object
...
info:
  contact: {}
  description: 'Personal finance tracking API. All /api/* endpoints require either
    a session cookie (browser) or an Authorization: Bearer <token> header (machine
    clients).'
  title: Pago API
  version: "1.0"
paths:
...
  /api/categories:
    get:
      parameters:
      - description: 'Transaction type: Income or Expense'
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.CategoriesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List categories
      tags:
      - transactions
...
  /api/stats:
    get:
      parameters:
      - description: Month in YYYY-MM (defaults to current month)
        in: query
        name: month
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.StatsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Monthly stats
      tags:
      - transactions
  /api/transactions:
    get:
      parameters:
      - description: Month in YYYY-MM (defaults to current month)
        in: query
        name: month
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.TransactionsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List transactions for a month
      tags:
      - transactions
...
  /api/transactions/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Transaction payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create transaction
      tags:
      - transactions
  /api/transactions/delete:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Transaction ID payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.DeleteTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete transaction
      tags:
      - transactions
...
securityDefinitions:
  BearerAuth:
    description: 'Provide the value as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"